+++ app/controllers/api/v2/bulk_controller.rb
@@ -7,9 +7,18 @@ class Api::V2::BulkController < Api::V2Controller
 
   def validate_and_transform_bulk_json_api_params
     @original_params = params
-    data_array = params.dig(:data) || []
 
-    bulk_data = data_array.map do |data|
+    @_params = ActionController::Parameters.new(
+      data: bulk_data_from_array_params(data_array_params),
+      action: params[:action],
+      controller: params[:controller]
+    )
+  end
+
+  def bulk_data_from_array_params(data_array_params)
+    data_array_params.map do |data|
+      require_id_for_all_data_objects unless data['data']['id']
+
       data[:action] = params[:action]
       data[:controller] = params[:controller]
 
@@ -17,11 +26,13 @@ class Api::V2::BulkController < Api::V2Controller
 
       JsonApiService.consume(params: data, context: self)
     end
+  end
 
-    @_params = ActionController::Parameters.new(
-      data: bulk_data,
-      action: params[:action],
-      controller: params[:controller]
-    )
+  def data_array_params
+    params.dig(:data) || []
+  end
+
+  def require_id_for_all_data_objects
+    raise Exceptions::BadRequestError, "An 'id' is required for every single object under /data"
   end
 end