+++ spec/acceptance/api/v2/tasks_spec.rb
@@ -23,18 +23,6 @@ resource 'Tasks' do
     build_data(new_task, account_list_id: user.account_lists.first.uuid)
   end
 
-  let(:bulk_update_form_data) do
-    [
-      {
-        data: {
-          type: resource_type,
-          id: task.uuid,
-          attributes: new_task
-        }
-      }
-    ]
-  end
-
   let(:resource_attributes) do
     %w(
       activity_type
@@ -189,25 +177,6 @@ resource 'Tasks' do
         do_request data: form_data
         expect(resource_object['subject']).to eq new_task['subject']
         expect(response_status).to eq 200
-      end
-    end
-
-    put '/api/v2/tasks/bulk' do
-      with_options scope: :data do
-        parameter 'id', 'Each member of the array must contain the id of the contact being updated'
-        parameter 'attributes', 'Each member of the array must contain an object with the attributes that must be updated'
-      end
-
-      response_field 'data',
-                     'List of Task objects that have been successfully updated and list of errors related to Task objects that were not updated successfully',
-                     'Type' => 'Array[Object]'
-
-      example 'Bulk update tasks', document: documentation_scope do
-        explanation 'Bulk Update a list of Tasks with an array of objects containing the ID and updated attributes'
-        do_request data: bulk_update_form_data
-
-        expect(response_status).to eq(200), invalid_status_detail
-        expect(json_response.first['data']['attributes']['name']).to eq new_task['name']
       end
     end